requirements/base.txt
decorator==4.3.0
Django==2.0.6
django-cors-headers==2.2.0
django-extensions==2.0.7
djangorestframework==3.8.2
djangorestframework-expander==0.2.3
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
django-filter==2.0.0
python-decouple==3.1
psycopg2-binary==2.7.4
tablib==0.12.1
Pygments==2.2.0
requests==2.18.4
pytz==2018.4
ipdb==0.11
Werkzeug==0.14.1
django-environ==0.4.4
Unidecode==0.4.9
